+++ Standard Graph Types/SimpleDirectedGraph.h
@@ -20,9 +20,9 @@ public:
     ~SimpleDirectedGraph();
 
     //ACCESSORS
-    size_t netDegree(const Vertex<T> &v) const;
-    size_t inDegree(const Vertex<T> &v) const;
-    size_t outDegree(const Vertex<T> &v) const;
+    unsigned netDegree(const Vertex<T> &v) const;
+    unsigned inDegree(const Vertex<T> &v) const;
+    unsigned outDegree(const Vertex<T> &v) const;
 
     //MUTATORS
     pair<Edge<T>*, bool> addEdge(Edge<T> &e);
